backend/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
SQLAlchemy==1.3.24
psycopg2-binary==2.8.6
python-dotenv==0.19.0
pydantic==1.8.2
pandas==1.3.5
apache-flink==1.15.0
neo4j==5.14.0
kafka-python==2.0.2
xlsxwriter==3.0.3
python-multipart==0.0.5
openpyxl==3.1.2
xlrd==2.0.1
confluent-kafka==2.3.0
scikit-learn==1.0.2
matplotlib==3.5.3
seaborn==0.11.2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
opencv-python==4.8.0.76
click==8.0.4
h11==0.14.0
asgiref==3.7.2
typing_extensions==4.5.0
starlette==0.14.2
numpy==1.21.6
pytz==2021.3
python-dateutil==2.8.2
six==1.16.0
websockets==11.0.3 